requirements.txt
python-dotenv
rich
InquirerPy
requests